\name{CPD-class}
\Rdversion{1.1}
\docType{class}
\alias{CPD-class}
\alias{show,CPD-method}

\title{Class \code{"CPD"}}

\description{
  Class union of \code{DEFLP}, \code{DEFQP} and \code{DEFNL}.
}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Methods}{
No methods defined with class "CPD" in the signature.
}

\keyword{classes}
