\name{LNLV-class}
\Rdversion{1.1}
\docType{class}
\alias{LNLV-class}

\title{Class \code{"LNLV"}}

\description{
  Class union of \code{NNOV} and \code{NLFV}.
}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Methods}{
No methods defined with class "LNLV" in the signature.
}

\keyword{classes}
