\name{CPG}
\alias{CPG}
\alias{PDV}
\alias{DQP}

\title{
	Rcpp module: CPG
}

\description{
  Module for defining and solving convex programs.
}

\details{
  The module contains the following items:
  classes: \describe{
    \item{PDV}{Class for primal/dual variables.}
    \item{PDV}{Class for definition of quadratic programs.}
  }
}
    
\keyword{datasets}
