\name{CTL}
\alias{CTL}
\alias{CTRL}

\title{
	Rcpp module: CTL
}

\description{
  Module for defining options used in optimization routines.
}

\details{
  The module contains the following items:
  classes: \describe{
    \item{CTRL}{Class for control options in optimizations.}
  }
}
    
\keyword{datasets}
