\name{NLF}
\alias{NLF}
\alias{NLFV}
\alias{NLFS}

\title{
	Rcpp module: NLF
}

\description{
  Module for creating objects with respect to nonlinear constraints
  and operations for this kind of objects.
}

\details{
  The module contains the following items:
  classes: \describe{
    \item{NLFV}{Class for variables with respect to nonlinear
      constraints.}
    \item{NLFS}{Class for NT-scaling and Lagrange-Multipliers.}
  }
}
    
\keyword{datasets}
