\name{NNO}
\alias{NNO}
\alias{NNOV}
\alias{NNOS}
\alias{NNOC}

\title{
	Rcpp module: NNO
}

\description{
  Module for creating objects with respect to nonlinear constraints
  and operations for this kind of objects.
}

\details{
  The module contains the following items:
  classes: \describe{
    \item{NNOV}{Class for variables with respect to nonlinear
      constraints.}
    \item{NNOS}{Class for NT-scaling and Lagrange-Multipliers.}
    \item{NNOC}{Class for linear constraints.}
  }
}
    
\keyword{datasets}
