\name{PSD}
\alias{PSD}
\alias{PSDV}

\title{
	Rcpp module: PSD
}

\description{
  Module for creating objects with respect to nonlinear constraints
  and operations for this kind of objects.
}

\details{
  The module contains the following items:
  classes: \describe{
    \item{PSDV}{Class for variables with respect to nonlinear
      constraints.}
  }
}
    
\keyword{datasets}
