\name{Rcpp_NNOS-class}
\Rdversion{1.1}
\docType{class}
\alias{Rcpp_NNOS-class}

\title{Class \code{"Rcpp_NNOS"}}

\description{
  Class for variables that are related to linear constraints.
}

\section{Extends}{
Class \code{"\linkS4class{C++Object}"}, directly.
All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.
}
       
\section{Fields}{
  \describe{
    \item{\code{d}:}{Object of class \code{activeBindingFunction} ~~ }
    \item{\code{di}:}{Object of class \code{activeBindingFunction} ~~ }
    \item{\code{lambda}:}{Object of class \code{activeBindingFunction} ~~ }
  }
}

\section{Class-Based Methods}{
  \describe{
    \item{\code{initialize(...)}:}{ ~~ }
    \item{\code{finalize()}:}{ ~~ }
  }
}

\examples{
showClass("Rcpp_NNOS")
}
\keyword{classes}
