\name{Rcpp_PSDS-class}
\Rdversion{1.1}
\docType{class}
\alias{Rcpp_PSDS-class}

\title{Class \code{"Rcpp_PSDS"}}

\description{
  Class for scalings applied to variables that are related to positive
  semidefinite cone constraints. 
}

\section{Extends}{
Class \code{"\linkS4class{C++Object}"}, directly.
All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.
}
       
\section{Fields}{
  \describe{
    \item{\code{r}:}{Object of class \code{activeBindingFunction} ~~ }
    \item{\code{rti}:}{Object of class \code{activeBindingFunction} ~~ }
    \item{\code{lambda}:}{Object of class \code{activeBindingFunction} ~~ }
  }
}

\section{Class-Based Methods}{
  \describe{
    \item{\code{initialize(...)}:}{ ~~ }
    \item{\code{finalize()}:}{ ~~ }
  }
}

\examples{
showClass("Rcpp_PSDS")
}
\keyword{classes}
