\name{Rcpp_SOCS-class}
\Rdversion{1.1}
\docType{class}
\alias{Rcpp_SOCS-class}

\title{Class \code{"Rcpp_SOCS"}}

\description{
  Class for variables that are related to second-order cone constraints.
}

\section{Extends}{
Class \code{"\linkS4class{C++Object}"}, directly.
All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.
}
       
\section{Fields}{
  \describe{
    \item{\code{v}:}{Object of class \code{activeBindingFunction} ~~ }
    \item{\code{beta}:}{Object of class \code{activeBindingFunction} ~~ }
    \item{\code{lambda}:}{Object of class \code{activeBindingFunction} ~~ }
  }
}

\section{Class-Based Methods}{
  \describe{
    \item{\code{initialize(...)}:}{ ~~ }
    \item{\code{finalize()}:}{ ~~ }
  }
}

\examples{
showClass("Rcpp_SOCS")
}
\keyword{classes}
