\name{Rcpp_SOCV-class}
\Rdversion{1.1}
\docType{class}
\alias{Rcpp_SOCV-class}

\title{Class \code{"Rcpp_SOCV"}}

\description{
  Class for variables that are related to second-order cone constraints.
}

\section{Extends}{
Class \code{"\linkS4class{C++Object}"}, directly.
All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.
}
       
\section{Fields}{
  \describe{
    \item{\code{dims}:}{Object of class \code{activeBindingFunction} ~~ }
    \item{\code{u}:}{Object of class \code{activeBindingFunction} ~~ }
  }
}

\section{Class-Based Methods}{
  \describe{
    \item{\code{initialize(...)}:}{ ~~ }
    \item{\code{finalize()}:}{ ~~ }
    \item{\code{uone(...)}:}{ ~~ }
    \item{\code{jdot(...)}:}{ ~~ }
    \item{\code{udot(...)}:}{ ~~ }
    \item{\code{uprd(...)}:}{ ~~ }
    \item{\code{uinv(...)}:}{ ~~ }
    \item{\code{umsa(...)}:}{ ~~ }
    \item{\code{umss(...)}:}{ ~~ }
  }
}

\examples{
showClass("Rcpp_SOCV")
}
\keyword{classes}
