\name{CPD-class}
\Rdversion{1.1}
\docType{class}
\alias{CPD-class}

\title{Class \code{"CPD"}}

\description{
  Class union of \code{Rcpp_DLP}, \code{Rcpp_DQP}, \code{Rcpp_DCP}
  and \code{Rcpp_DNL}.
}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Methods}{
No methods defined with class "CPD" in the signature.
}

\keyword{classes}
