\name{CPG}
\alias{CTRL}
\alias{CPG}
\alias{CONEC}
\alias{PDV}
\alias{DLP}
\alias{DQP}
\alias{CPS}

\title{
	Rcpp module: CPG
}

\description{
  Module for defining and solving convex programs.
}

\details{
  The module contains the following items:
  classes: \describe{
    \item{CTRL}{Class for control parameters used in optimizations.}
    \item{CONEC}{Class for inequality (cone) constraints.}
    \item{PDV}{Class for primal/dual variables.}
    \item{DLP}{Class for definition of linear programs.}
    \item{DQP}{Class for definition of quadratic programs.}
    \item{CPS}{Class for solution of convex programs.}
  }
}
    
\keyword{datasets}
