\name{PSD}
\alias{PSD}
\alias{PSDV}
\alias{PSDS}
\alias{PSDC}

\title{
	Rcpp module: PSD
}

\description{
  Module for creating objects with respect to nonlinear constraints
  and operations for this kind of objects.
}

\details{
  The module contains the following items:
  classes: \describe{
    \item{PSDV}{Class for variables with respect to nonlinear
      constraints.}
    \item{PSDS}{Class for NT-scaling and Lagrange-Multipliers.}
    \item{PSDC}{Class for positive semidefinite constraints.}
  }
}
    
\keyword{datasets}
