\name{SOC}
\alias{SOC}
\alias{SOCV}
\alias{SOCS}

\title{
	Rcpp module: SOC
}

\description{
  Module for creating objects with respect to nonlinear constraints
  and operations for this kind of objects.
}

\details{
  The module contains the following items:
  classes: \describe{
    \item{SOCV}{Class for variables with respect to nonlinear
      constraints.}
    \item{SOCS}{Class for NT-scaling and Lagrange-Multipliers.}
  }
}
    
\keyword{datasets}
