\name{Rcpp_CTRL-class}
\Rdversion{1.1}
\docType{class}
\alias{Rcpp_CTRL-class}

\title{Class \code{"Rcpp_CTRL"}}

\description{
  Class for control options used in optimization routines.
}

\section{Extends}{
Class \code{"\linkS4class{C++Object}"}, directly.
All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.
}
       
\section{Fields}{
  \describe{
    \item{\code{ctrlparams}:}{Object of class
      \code{activeBindingFunction}: List of control parameters.}
  }
}

\examples{
showClass("Rcpp_CTRL")
}
\keyword{classes}
